JS
Got stuck on why the links for the experience section could not be clicked. Thought it was the lack of an event handler because the dom elements are loaded in, but it was a silly z-index problem. Spent way too much time on that part, will check first if ever happens again.